requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
sentence-transformers==2.7.0
faiss-cpu==1.13.2
openai==1.3.5
slowapi==0.1.9
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0